+++ src/components/LogInPage.css
@@ -1,0 +1,284 @@
+@import url('https://fonts.googleapis.com/css2?family=Frank+Ruhl+Libre:wght@400;500;700&display=swap');
+@import url('https://fonts.googleapis.com/css2?family=Alegreya+SC:wght@400;500;700&display=swap');
+@import url('https://fonts.googleapis.com/css2?family=Source+Serif+Pro:wght@200;300;400&display=swap');
+@import url('https://fonts.googleapis.com/css2?family=Gothic+A1:wght@400;500;700&display=swap');
+
+.back-box {
+    position: absolute;
+    top: 50%;
+    left: 50%;
+    transform: translate(-51%, -55%);
+    background: rgba(39, 43, 44, 10);
+    width: 509px;
+    height: 319px;
+    border-radius: 10px;
+    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.3);
+    border: 2px solid white;;
+}
+
+.logo {
+    position: absolute;
+    top: 50%;
+    left: 28%;
+    transform: translate(-50%, -50%);
+    width: 20% !important;
+}
+
+.title-box {
+    color: white;
+    position: absolute;
+    top: 52%;
+    left: 60%;
+    transform: translate(-50%, -50%);
+}
+
+.title {
+    font-size: 48px !important;
+    letter-spacing: 8px !important;
+    font-family: 'Frank Ruhl Libre', serif;
+    font-weight: 500;
+}
+
+.minititle {
+    font-size: 12px !important;
+    letter-spacing: 10px !important;
+    font-family: 'Alegreya SC', serif;
+}
+
+/* 스크롤 숨기기 */
+body, html {
+    overflow: hidden; /* 스크롤 비활성화 */
+  }
+
+.start-page {
+    width: 100vw;
+    height: 100vh;
+    background-image: url('../../public/images/InitPage.png'); /* 배경 이미지 */
+    background-size: cover; 
+    background-position: center;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    cursor: pointer;
+  }
+  
+  .content {
+    text-align: center;
+  }
+  
+  .start-text {
+    width: 100%;
+    height: auto;
+    max-width: 1000px;
+    max-height: 1000px;
+    object-fit: contain;
+    display: block;
+    margin: 0 auto;
+    position: relative;
+    top: -70px;
+    transition: opacity 0.6s ease-in-out, transform 2s ease-in-out;
+    opacity: 1;
+  }
+  
+  .start-text.fade-out {
+    opacity: 0;
+  }
+
+.login-box {
+    position: absolute;
+    top: 50%;
+    left: 50%;
+    transform: translate(-50%, -50%);
+    background-color: rgba(39, 43, 44, 0.8);
+    width: 600px;
+    height: 400px;
+    border-radius: 10px;
+    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.3);
+    border: 2px solid white;
+}
+
+.login-header {
+    display: flex;
+    align-items: center;
+    color: white;
+    padding: 15px;
+}
+
+.log-logo {
+    width: 14%;
+}
+
+.log-title-contain {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    margin-left: 10px;
+    padding-top: 12px;
+} 
+
+.log-minititle {
+    font-size: 12px;
+    font-family: 'Alegreya SC', serif;
+    letter-spacing: 6px;
+}
+
+.log-title {
+    font-size: 35px;
+    font-family: 'Frank Ruhl Libre', serif;
+    font-weight: 600;
+    letter-spacing: 6px;
+}
+
+.log-subtitle {
+    position: absolute;
+    left: 73%;
+    font-family: 'Gothic A1', serif;
+    font-weight: 300;
+}
+
+
+.login-body {
+    position: absolute;
+    display: flex;
+    flex-direction: column;
+    width: 55%;
+    top: 37%;
+    left: 42%;
+    gap: 20px;
+}
+
+.login-input {
+    height: 50px;
+    border-radius: 5px;
+    font-family: 'Source Serif Pro', serif;
+    background-color: rgba(72, 75, 75, 0.7);
+    box-shadow: inset 0 4px 8px rgba(0, 0, 0, 0.3);
+    padding: 10px;
+}
+
+.login-button {
+    position: absolute;
+    bottom: 4%;
+    color: white;
+    font-family: 'Source Serif Pro', serif;
+    font-size: 20px;
+    letter-spacing: 2px;
+    font-weight: 300;
+    width: 598px;
+    border-top: 1px solid white;
+    padding-top: 9px;
+}
+
+.log-profile {
+    position: absolute;
+    top: 30%;
+    left: 5%;
+    width: 30%;
+} 
+
+.signupPage {
+    position: absolute;
+    top: 110%;
+    left: 50%;
+    transform: translate(-50%, -50%);
+    color: white;
+    font-family: 'Gothic A1', serif;
+    letter-spacing: 4px;
+}
+
+.centered-image-container {
+    position: absolute;
+    top: 50%;
+    left: 50%;
+    transform: translate(-50%, -50%);
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    flex-direction: column;
+    z-index: 10;
+    width: 100%;
+    height: 100%;
+  }
+  
+  .wallet-image {
+    position: absolute;
+    top: 50%;
+    left: 50%;
+    transform: translate(-50%, -50%);
+    width: 600px;
+    height: auto;
+    object-fit: contain;
+    opacity: 1;
+    transition: transform 2s ease-in-out, opacity 3s ease-in-out;
+  }
+  
+  .cutting-wallet {
+    position: absolute;
+    top: 50%;
+    left: 50%;
+    transform: translate(-49.5%, -40.6%);
+    width: 613px;
+    height: auto;
+    object-fit: contain;
+    opacity: 1;
+    transition: transform 2s ease-in-out, opacity 3s ease-in-out;
+  }
+  
+  .wallet-image.move-down,
+  .cutting-wallet.move-down {
+    transform: translate(-50%, 150%);
+    opacity: 0;
+  }
+  
+  .idcard-container {
+      perspective: 1000px; /* 3D 효과 */
+  }
+  
+  
+  .idcard-container .idcard-front,
+  .idcard-container .idcard-back {
+    position: absolute;
+    width: 100%;
+    height: 100%;
+    backface-visibility: hidden; /* 뒷면 숨김 */
+    transition: transform 1s ease-in-out;
+  }
+  
+  .idcard-container .idcard-back {
+    transform: rotateY(180deg);
+  }
+  
+  .idcard-container.flip .idcard-front {
+    transform: rotateY(-180deg);
+  }
+  
+  .idcard-container.flip .idcard-back {
+    transform: rotateY(0deg);
+  }
+  
+  
+
+  
+
+  .zoom-background {
+    width: 100vw;
+    height: 100vh;
+    background-image: url('../../images/InitPage.png'); 
+    background-size: cover;
+    background-position: center;
+    animation: zoomIn 2s ease-in-out forwards; 
+  }
+  
+  @keyframes zoomIn {
+    0% {
+      transform: scale(1);
+    }
+    100% {
+      transform: scale(1.25); 
+    }
+  }
+  
+  /* 스크롤 숨기기 */
+body, html {
+  overflow: hidden; /* 스크롤 비활성화 */
+}
