--- src/components/LoginBox.css
@@ -1,106 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=Frank+Ruhl+Libre:wght@400;500;700&display=swap');
-@import url('https://fonts.googleapis.com/css2?family=Alegreya+SC:wght@400;500;700&display=swap');
-@import url('https://fonts.googleapis.com/css2?family=Source+Serif+Pro:wght@200;300;400&display=swap');
-@import url('https://fonts.googleapis.com/css2?family=Gothic+A1:wght@400;500;700&display=swap');
-
-.login-box {
-    position: absolute;
-    top: 50%;
-    left: 50%;
-    transform: translate(-50%, -50%);
-    background-color: rgba(39, 43, 44, 0.8);
-    width: 600px;
-    height: 400px;
-    border-radius: 10px;
-    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.3);
-    border: 2px solid white;
-}
-
-.login-header {
-    display: flex;
-    align-items: center;
-    color: white;
-    padding: 15px;
-}
-
-.log-logo {
-    width: 14%;
-}
-
-.log-title-contain {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    margin-left: 10px;
-    padding-top: 12px;
-} 
-
-.log-minititle {
-    font-size: 12px;
-    font-family: 'Alegreya SC', serif;
-    letter-spacing: 6px;
-}
-
-.log-title {
-    font-size: 35px;
-    font-family: 'Frank Ruhl Libre', serif;
-    font-weight: 600;
-    letter-spacing: 6px;
-}
-
-.log-subtitle {
-    position: absolute;
-    left: 73%;
-    font-family: 'Gothic A1', serif;
-    font-weight: 300;
-}
-
-
-.login-body {
-    position: absolute;
-    display: flex;
-    flex-direction: column;
-    width: 55%;
-    top: 37%;
-    left: 42%;
-    gap: 20px;
-}
-
-.login-input {
-    height: 50px;
-    border-radius: 5px;
-    font-family: 'Source Serif Pro', serif;
-    background-color: rgba(72, 75, 75, 0.7);
-    box-shadow: inset 0 4px 8px rgba(0, 0, 0, 0.3);
-    padding: 10px;
-}
-
-.login-button {
-    position: absolute;
-    bottom: 4%;
-    color: white;
-    font-family: 'Source Serif Pro', serif;
-    font-size: 20px;
-    letter-spacing: 2px;
-    font-weight: 300;
-    width: 598px;
-    border-top: 1px solid white;
-    padding-top: 9px;
-}
-
-.log-profile {
-    position: absolute;
-    top: 30%;
-    left: 5%;
-    width: 30%;
-} 
-
-.signupPage {
-    position: absolute;
-    top: 110%;
-    left: 50%;
-    transform: translate(-50%, -50%);
-    color: white;
-    font-family: 'Gothic A1', serif;
-    letter-spacing: 4px;
-}
