--- src/components/StartPage.css
@@ -1,38 +1,0 @@
-.start-page {
-  width: 100vw;
-  height: 100vh;
-  background-image: url('../../images/InitPage.png'); /* 배경 이미지 */
-  background-size: cover; /* 배경 크기 */
-  background-position: center; /* 배경 위치 */
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  cursor: pointer; /* 클릭 가능 커서 */
-}
-
-.content {
-  text-align: center;
-}
-
-.start-text {
-  width: 100%;
-  height: auto;
-  max-width: 1000px;
-  max-height: 1000px;
-  object-fit: contain;
-  display: block;
-  margin: 0 auto;
-  position: relative;
-  top: -70px;
-  transition: opacity 0.6s ease-in-out, transform 2s ease-in-out;
-  opacity: 1;
-}
-
-.start-text.fade-out {
-  opacity: 0;
-}
-
-/* 스크롤 숨기기 */
-body, html {
-  overflow: hidden; /* 스크롤 비활성화 */
-}
