--- src/components/ZoomBackground.css
@@ -1,22 +1,0 @@
-.zoom-background {
-    width: 100vw;
-    height: 100vh;
-    background-image: url('../../images/InitPage.png'); 
-    background-size: cover;
-    background-position: center;
-    animation: zoomIn 2s ease-in-out forwards; 
-  }
-  
-  @keyframes zoomIn {
-    0% {
-      transform: scale(1);
-    }
-    100% {
-      transform: scale(1.25); 
-    }
-  }
-  
-  /* 스크롤 숨기기 */
-body, html {
-  overflow: hidden; /* 스크롤 비활성화 */
-}
