--- src/pages/ZoomBackground.tsx
@@ -1,8 +1,0 @@
-import React from "react";
-import '../components/ZoomBackground.css';
-
-const ZoomBackground: React.FC = () => {
-  return <div className="zoom-background"></div>;
-};
-
-export default ZoomBackground;
